[#160021786] create query to create database table, refactor create a user and login method
