+++ .bablerc
@@ -1,5 +1,8 @@
 {
 "presets":[
-"airbnb"
+"es2015", "airbnb"
+],
+"plugins":[
+    "babel-plugin-inline-import"
 ]
 }
