+++ .travis.yml
@@ -11,5 +11,6 @@ script:
     - npm run coveralls
 before_script:
   - npm install codeclimate-test-reporter istanbul -g
+  - npm run build
 after_script:
   - codeclimate-test-reporter < ./coverage/lcov.info
