--- es6/server/auth/check-auth.js
@@ -1,15 +1,0 @@
-import jwt from 'jsonwebtoken';
-
-export default (req, res, next) => {
-    try{
-        const token = req.headers.authorization;
-        const decoded = jwt.verify(token, process.env.JWT_KEY);
-        req.userData = decoded;
-        next();
-    }
-    catch(error){
-        return res.status(401).json({
-            message: 'Authentication Failed!'
-        });
-    }
-}
