--- es6/server/db/index.js
@@ -1,3 +1,0 @@
-require('dotenv').config();
-const pgp = require('pg-promise')();
-export const db = pgp('postgresql://CzA:King301094@localhost:5432/stackoverflow');
