+++ package.json
@@ -4,25 +4,28 @@
   "description": "Two week bootcamp project",
   "main": "server.js",
   "scripts": {
-    "start": "node server/server.js",
-    "test": "mocha server/test",
-    "cover": "istanbul cover node_modules/mocha/bin/_mocha server/test",
+    "start": "npm run build && nodemon compiled/server.js",
+    "test": "mocha compiled/test",
+    "cover": "istanbul cover node_modules/mocha/bin/_mocha compiled/test",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls",
     "test-watch": "nodemon --exec \"npm test\"",
-    "build": "babel --presets airbnb es6/server -d server/",
-    "build2": "babel --presets airbnb es6/es6styles -d public/js"
+    "build": "babel --presets airbnb server -d compiled/"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "babel-core": "^6.26.3",
     "bcrypt": "^3.0.0",
     "body-parser": "^1.18.3",
     "dotenv": "^6.0.0",
+    "eslint-plugin-import": "^2.14.0",
     "express": "^4.16.3",
     "hbs": "^4.0.1",
     "jsonwebtoken": "^8.3.0",
+    "make-runnable": "^1.3.6",
     "morgan": "^1.9.0",
     "multer": "^1.3.1",
+    "nodemon": "^1.18.3",
     "pg": "^7.4.3",
     "pg-promise": "^8.4.6",
     "request": "^2.88.0"
@@ -37,7 +40,6 @@
     "eslint": "^5.3.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-react": "^7.11.1",
     "istanbul": "^0.4.5",
