+++ server/auth/check-auth.js
@@ -1,24 +1,15 @@
-'use strict';
+import jwt from 'jsonwebtoken';
 
-Object.defineProperty(exports, "__esModule", {
-    value: true
-});
-
-var _jsonwebtoken = require('jsonwebtoken');
-
-var _jsonwebtoken2 = _interopRequireDefault(_jsonwebtoken);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-exports['default'] = function (req, res, next) {
-    try {
-        var token = req.headers.authorization;
-        var decoded = _jsonwebtoken2['default'].verify(token, process.env.JWT_KEY);
+export default (req, res, next) => {
+    try{
+        const token = req.headers.authorization;
+        const decoded = jwt.verify(token, process.env.JWT_KEY);
         req.userData = decoded;
         next();
-    } catch (error) {
+    }
+    catch(error){
         return res.status(401).json({
             message: 'Authentication Failed!'
         });
     }
-};
+}
