+++ server/controllers/user.js
@@ -1,113 +1,103 @@
-'use strict';
+import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
+import { db } from '../db';
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports.User = undefined;
 
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
+export class User {
+	// create a new user
+	createUser(req, res) {
+		const firstName = req.body.firstName;
+		const lastName = req.body.lastName;
+		const fullName = `${firstName} ${lastName}`;
+		const username = req.body.username;
+		const emailAdd = req.body.email;
+		const pwd = req.body.password;
 
-var _bcrypt = require('bcrypt');
+		const email = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(emailAdd);
+		const password = (/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[#$^+=!*()@%&]).{6,10}$/g).test(pwd) && pwd.length >= 6;
+		const first = typeof firstName === 'string' && firstName.length > 1 && (/^[a-zA-Z]+(([' -][a-zA-Z ])?[a-zA-Z]*)*$/g).test(firstName);
+		const last = typeof lastName === 'string' && lastName.length > 1 && (/^[a-zA-Z]+(([' -][a-zA-Z ])?[a-zA-Z]*)*$/g).test(lastName);
 
-var _bcrypt2 = _interopRequireDefault(_bcrypt);
-
-var _jsonwebtoken = require('jsonwebtoken');
-
-var _jsonwebtoken2 = _interopRequireDefault(_jsonwebtoken);
-
-var _db = require('../db');
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-var User = exports.User = function () {
-	function User() {
-		_classCallCheck(this, User);
+		if (first && last && username && email && password) {
+			db.any('SELECT * FROM users WHERE email = $1 OR username = $2', [emailAdd, username])
+			.then( user => {
+				if (user.length === 0) {
+						const hash = bcrypt.hashSync(pwd, 10);
+						return db.any('INSERT INTO users (fullname, username, email, password) VALUES ($1, $2, $3, $4)', [fullName, username,emailAdd, hash])
+							.then(() => {
+								const token = jwt.sign({
+									emailAdd 
+								},process.env.JWT_KEY,{
+									expiresIn: '1h'
+								});
+								res.status(201).json({ message: 'user created', token })
+							}
+							)
+							.catch((error) => res.status(500).json({error}));
+				}
+				if (user[0].email === emailAdd) {
+					return res.status(409).json({ message: 'a user with this email already exists' });
+				}
+				if (user[0].username === username) {
+					return res.status(409).json({ message: 'a user with this username already exists' });
+				}
+			})
+			.catch((error) => {
+			res.status(500).json(error);
+		});
+		}
+		if(!firstName || first === false){
+			return res.status(400).json({ message: 'first name is empty or an invalid format' });
+		}
+		if(!lastName || last === false){
+			return res.status(400).json({ message: 'last name is empty or an invalid format' });
+		}
+		if(!username){
+			return res.status(400).json({ message: 'username is empty or an invalid format' });
+		}
+		if(!emailAdd || email === false){
+			return res.status(400).json({ message: 'email is empty or an invalid format' });
+		} 
+		if(!pwd || password === false){
+			return res.status(400).json({ message: 'password is empty or an invalid format' });
+		}
 	}
 
-	_createClass(User, [{
-		key: 'createUser',
-
-		// create a new user
-		value: function () {
-			function createUser(req, res) {
-				_db.db.any('SELECT * FROM users WHERE email = $1', [req.body.email]).then(function (user) {
-					if (user.length === 0) {
-						var email = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(req.body.email);
-						var password = /[a-zA-Z]/.test(req.body.password) && /[\W\d]/.test(req.body.password) && req.body.password.length >= 6;
-						if (email && password) {
-							var hash = _bcrypt2['default'].hashSync(req.body.password, 10);
-							return _db.db.any('INSERT INTO users (email, password) VALUES ($1, $2)', [req.body.email, hash]).then(function () {
-								return res.status(201).json({ message: 'user created' });
-							})['catch'](function (error) {
-								return res.status(400).json({ error: error });
-							});
-						}
-						if (!email) {
-							return res.status(400).json({ message: 'email is an invalid format' });
-						}
-						if (!password) {
-							return res.status(400).json({ message: 'password is an invalid format' });
-						}
-					}
-					return res.status(409).json({ message: 'User mail exists' });
-				})['catch'](function (error) {
-					res.status(400).json(error);
-				});
+	// login to an account
+	login (req, res) {
+		const email = req.body.email;
+		const pwd = req.body.password;
+		db.any('SELECT * FROM users WHERE email = $1', [email])
+		.then( user => {
+			if (user.length < 1) {
+				return res.status(401).json({message: 'invalid user!'});
 			}
+			bcrypt.compare(pwd, user[0].password, (err, result) => {
+				if (err) {
+					return res.status(401).json({message: 'invalid user!'});
+				}
+				if (result) {
+					const token = jwt.sign({
+						email: user[0].email 
+					},process.env.JWT_KEY,{
+						expiresIn: '1h'
+					});					
+					return res.status(200).json({ message: 'authentication successful!', token });
+				}
+				return res.status(401).json({ message: 'email and password do not match' });
+			});
+		})
+		.catch(
+			error => res.status(500).json({ error })
+		);
+	}
 
-			return createUser;
-		}()
-
-		// login to an account
-
-	}, {
-		key: 'login',
-		value: function () {
-			function login(req, res) {
-				_db.db.any('SELECT * FROM users WHERE email = $1', [req.body.email]).then(function (user) {
-					if (user.length < 1) {
-						return res.status(401).json({ message: 'Authentication failed!' });
-					}
-					_bcrypt2['default'].compare(req.body.password, user[0].password, function (err, result) {
-						if (err) {
-							return res.status(401).json({ message: 'Authentication failed!' });
-						}
-						if (result) {
-							var token = _jsonwebtoken2['default'].sign({
-								email: user[0].email,
-								userid: user[0].id
-							}, process.env.JWT_KEY, {
-								expiresIn: "1h"
-							});
-
-							return res.status(200).json({ message: 'Authentication successful!' });
-						}
-						return res.status(401).json({ message: 'Authentication failed!' });
-					});
-				})['catch']();
-			}
-
-			return login;
-		}()
-
-		// fetch all users
-
-	}, {
-		key: 'fetchUsers',
-		value: function () {
-			function fetchUsers(req, res) {
-				_db.db.any('SELECT * FROM users').then(function (users) {
-					return res.status(200).json({ users: users });
-				})['catch'](function (error) {
-					return res.status(400).json({ error: error });
-				});
-			}
-
-			return fetchUsers;
-		}()
-	}]);
-
-	return User;
-}();
+	// fetch all users
+	fetchUsers(req, res) {
+		db.any('SELECT * FROM users').then(
+			users => res.status(200).json({ users })
+		).catch(
+			error => res.status(500).json({ error })
+		);
+	}
+}
