+++ server/db/questions.json
@@ -1,57 +1,1 @@
-[
-  {
-    "questionId": "question1",
-    "title": "How to use Javascript",
-    "content": "I'm new to javascript. How do I use it to hide elements",
-    "answer": [
-      "With JQuery use $('.class or #id').hide",
-      "select element with let item =  document.getElementById. Then item.style.display = 'none'"
-    ]
-  },
-  {
-    "questionId": "question5",
-    "title": "How to use stackOverflowLite",
-    "content": "I'm new to stackOverflowLite. How do I ask a question",
-    "answer": [
-      "On the right side of the page click on the pencil button",
-      "Look for the post a question button dummy"
-    ]
-  },
-  {
-    "questionId": "question6",
-    "title": "How do I get into Andela?",
-    "content": "I've been hearing about Andela. How do I apply to their fellowship program?",
-    "answer": []
-  },
-  {
-    "questionId": "question7",
-    "title": "How to use babel to compile es6 code",
-    "content": "visit their website and follow instructions",
-    "answer": []
-  },
-  {
-    "questionId": "question2",
-    "title": "How to use JSON",
-    "content": "I'm new to JSON. How do I parse a JSON file",
-    "answer": [
-      "use JSON.parse(jsonFile)"
-    ]
-  },
-  {
-    "questionId": "question3",
-    "title": "How to use Ruby on Rails",
-    "content": "How do I run a for loop in Ruby on rails",
-    "answer": [
-      "I dont know2",
-      "I dont know"
-    ]
-  },
-  {
-    "questionId": "question4",
-    "title": "How to use Python",
-    "content": "I'm new to Python. How do initialize a variable",
-    "answer": [
-      "int i = 8"
-    ]
-  }
-]
+[]
